Regenerated Interop dlls with larger PROPVARIANT struct
PROPVARIANT struct contains a few fields and an union. COM Interop
thinks that the union is 8 bytes large, but on x64 it is probably larger,
as methods working with this in native code overwrote C# variables next to
it in memory. I suspect that the correct size of this struct should be 24,
but as I was unable to verify this, I explicitly set the size to 32.
As the struct is never passed by value, this size discrepancy shouldn't
be a problem. (And if it was, it would mean even bigger problems, as the
size changes between 32 and 64-bit.)

The struct is defined in both PDTypes and PDApi dlls, so both have
been updated.